separate strong reference test from implementation tests
